Refactor repository classes to remove the type (file vs. dynamodb) as these are now in different package locations.
