+++ src/main/java/gov/cdc/izgateway/xform/repository/file/AccessControlRepository.java
@@ -1,22 +1,22 @@
 package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import gov.cdc.izgateway.xform.model.OperationPreconditionField;
+import gov.cdc.izgateway.xform.model.AccessControl;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 import java.util.LinkedHashSet;
 
 @Repository
-public class OperationPreconditionFieldFileRepository extends GenericFileRepository<OperationPreconditionField> {
+public class AccessControlRepository extends GenericFileRepository<AccessControl> {
 
-    @Value("${xform.configurations.operation-precondition-fields}")
+    @Value("${xform.configurations.access-control}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
     @Override
-    protected TypeReference<LinkedHashSet<OperationPreconditionField>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<AccessControl>> getTypeReference() {
         return new TypeReference<>() {};
     }
 }
