+++ src/main/java/gov/cdc/izgateway/xform/repository/file/GroupRoleMappingRepository.java
@@ -1,22 +1,22 @@
 package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import gov.cdc.izgateway.xform.model.AccessControl;
+import gov.cdc.izgateway.xform.model.GroupRoleMapping;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 import java.util.LinkedHashSet;
 
 @Repository
-public class AccessControlFileRepository extends GenericFileRepository<AccessControl> {
+public class GroupRoleMappingRepository extends GenericFileRepository<GroupRoleMapping> {
 
-    @Value("${xform.configurations.access-control}")
+    @Value("${xform.configurations.group-role-mapping}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
     @Override
-    protected TypeReference<LinkedHashSet<AccessControl>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<GroupRoleMapping>> getTypeReference() {
         return new TypeReference<>() {};
     }
 }
