+++ src/main/java/gov/cdc/izgateway/xform/repository/file/OperationPreconditionFieldRepository.java
@@ -1,22 +1,22 @@
 package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import gov.cdc.izgateway.xform.model.Solution;
+import gov.cdc.izgateway.xform.model.OperationPreconditionField;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 import java.util.LinkedHashSet;
 
 @Repository
-public class SolutionFileRepository extends GenericFileRepository<Solution> {
+public class OperationPreconditionFieldRepository extends GenericFileRepository<OperationPreconditionField> {
 
-    @Value("${xform.configurations.solutions}")
+    @Value("${xform.configurations.operation-precondition-fields}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
     @Override
-    protected TypeReference<LinkedHashSet<Solution>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<OperationPreconditionField>> getTypeReference() {
         return new TypeReference<>() {};
     }
 }
