+++ src/main/java/gov/cdc/izgateway/xform/repository/file/SolutionRepository.java
@@ -1,22 +1,22 @@
 package gov.cdc.izgateway.xform.repository.file;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import gov.cdc.izgateway.xform.model.User;
+import gov.cdc.izgateway.xform.model.Solution;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 import java.util.LinkedHashSet;
 
 @Repository
-public class UserFileRepository extends GenericFileRepository<User> {
+public class SolutionRepository extends GenericFileRepository<Solution> {
 
-    @Value("${xform.configurations.users}")
+    @Value("${xform.configurations.solutions}")
     protected void setFilePath(String filePath) {
         this.filePath = filePath;
     }
 
     @Override
-    protected TypeReference<LinkedHashSet<User>> getTypeReference() {
+    protected TypeReference<LinkedHashSet<Solution>> getTypeReference() {
         return new TypeReference<>() {};
     }
 }
