+++ testing/configuration/group-role-mapping.json
@@ -2,27 +2,27 @@
   "id" : "fe65c254-eeee-4d41-a81c-187847a10dd0",
   "groupName" : "Xform Administrators",
   "active" : true,
-  "roles" : [ "pipeline-reader", "pipeline-writer", "pipeline-deleter", "solution-reader", "solution-writer", "solution-deleter", "organization-reader", "organization-writer", "organization-deleter", "admin" ]
+  "roles" : [ "admin", "organization-deleter", "organization-reader", "organization-writer", "pipeline-deleter", "pipeline-reader", "pipeline-writer", "solution-deleter", "solution-reader", "solution-writer" ]
 }, {
   "id" : "fe65c254-bbbb-4d41-a81c-187847a10dd0",
   "groupName" : "Xform Business Analyst",
   "active" : true,
-  "roles" : [ "pipeline-reader", "pipeline-writer", "pipeline-deleter", "organization-reader" ]
+  "roles" : [ "organization-reader", "pipeline-deleter", "pipeline-reader", "pipeline-writer" ]
 }, {
   "id" : "a3c729dd-4f0e-48c7-a6a6-18c66f8cba93",
   "groupName" : "Xform Solutions Engineer",
   "active" : true,
-  "roles" : [ "pipeline-reader", "pipeline-writer", "pipeline-deleter", "solution-reader", "solution-writer", "solution-deleter", "organization-reader", "organization-writer" ]
+  "roles" : [ "organization-reader", "organization-writer", "pipeline-deleter", "pipeline-reader", "pipeline-writer", "solution-deleter", "solution-reader", "solution-writer" ]
 }, {
   "id" : "6c31247a-9195-4d23-b203-12dfb485ad10",
   "groupName" : "Xform Onboarding Staff",
   "active" : true,
-  "roles" : [ "pipeline-reader", "solution-reader", "organization-reader" ]
+  "roles" : [ "organization-reader", "pipeline-reader", "solution-reader" ]
 }, {
   "id" : "86eb1912-b46d-435e-85e5-64d63853ffc6",
   "groupName" : "Xform Operations Staff",
   "active" : true,
-  "roles" : [ "pipeline-reader", "pipeline-writer", "pipeline-deleter", "solution-reader", "solution-writer", "solution-deleter", "organization-reader", "organization-writer" ]
+  "roles" : [ "organization-reader", "organization-writer", "pipeline-deleter", "pipeline-reader", "pipeline-writer", "solution-deleter", "solution-reader", "solution-writer" ]
 }, {
   "id" : "7b23ee9e-ba1a-43f7-9290-2b6a3dccaf32",
   "groupName" : "Xform Sender",
